==> 2) Creating the TF-IDF Representation.ipynb <==
import pandas as pd
import numpy as np
from nltk.stem import PorterStemmer
from sklearn.feature_extraction.text import CountVectorizer
import nltk
from sklearn.feature_extraction.text import ENGLISH_STOP_WORDS
from nltk.stem import PorterStemmer
from wordcloud import WordCloud

companies = pd.read_csv('cleanedText.csv')

companies.head()

companies.shape

#GOAL

#1. Do stemming on the data
#2. Add the data over the years
#3. Create TF-IDF matrix manually

companies[companies['finalText'].isna()]
#have to drop these rows

companies = companies[~companies['finalText'].isna()]

companies.shape

#Tokenize final text, remove stop words and do stemming
companies['finalText'] = companies['finalText'].apply(nltk.word_tokenize)
companies.head()

def remove_stop_words(new_text):
    result = [w.lower() for w in new_text if w.lower() not in ENGLISH_STOP_WORDS]
    return result

companies['finalText'] = companies.finalText.apply(remove_stop_words)
companies.head()

stemmer = PorterStemmer()

def stem_words(tokens):
    return [stemmer.stem(w) for w in tokens]

companies['finalText'] = companies.finalText.apply(stem_words)
companies.head()

def join_tokens(tokens):
    return " ".join(tokens)

companies['finalText'] = companies.finalText.apply(join_tokens)
companies.head()

companies.to_csv('CompaniesPostStemming.csv')

#Add the data of the same company to create on long string of all filings over the 2 years
tickers = companies['ticker'].unique()
len(tickers)

count = 0
companies2 = pd.DataFrame(columns=['ticker', 'text'])
for i in tickers:
    temp = companies[companies['ticker'] == i]
    companies2.loc[count] = [i, temp['finalText'].str.cat(sep = ' ')]
    count+=1
    

companies2

companies2['count'] = [len(text.split(' ')) for text in companies2['text']]

companies2.head()

#Now we create the TF-IDF matrix manually 
word_vectorizer = CountVectorizer(token_pattern = '[a-zA-Z]+')

# Now use this to set-up the same type of dataset as before:
vectorizer = word_vectorizer.fit(companies2['text'])
word_bow_matrix = vectorizer.transform(companies2['text'])
count_vectorizer = pd.DataFrame(word_bow_matrix.toarray(), 
                           columns = word_vectorizer.get_feature_names_out())

count_vectorizer.head()

tf_matrix = count_vectorizer/count_vectorizer.sum(axis = 0)
tf_matrix.head()

idf_matrix = tf_matrix.apply(lambda row: (row!=0).sum(), axis = 0)
idf_matrix = np.log(len(tf_matrix)/idf_matrix)
idf_matrix.head()

tfidf_matrix = tf_matrix * idf_matrix
tfidf_matrix.head()

tfidf_matrix.shape, companies2.shape #496 companies

tfidf_matrix.to_csv('TF-IDF.csv')
