==> 6) Profit Pool.ipynb <==
import numpy as np
import pandas as pd
import yfinance as yf
import plotly.express as px

## Input
lookback: how many years to lookback  
x: measures in x axis  
y: measures in y axis  
sector: define specific sectors to plot (if input is an empty list, then plot all sectors)

# input
lookback = 3
x = 'Net Income'
y = 'Profit Margin'
sector = ['Industrials']

# run function
plot_pp(lookback, x, y, sector)

# function definition
def plot_pp(lookback, x, y, sector):
    # read data
    pp_df = pd.read_csv('pp_df.csv')
    # plot
    plot_df = pp_df[pp_df['Year'] > pp_df['Year'].max() - lookback].reset_index(drop=True)
    plot_df['Year'] = plot_df['Year'].astype(str)
    if (len(sector) == 0):
        for ind in plot_df['GICS Sector'].unique():
            fig = px.scatter(plot_df[plot_df['GICS Sector'] == ind], x = x, y = y, color = 'Year', hover_name= 'Symbol')
            fig.update_layout(title = 'Profit Pool of ' + ind,
                          xaxis_title = x,
                          yaxis_title = y,
                          title_x = 0.5,
                          plot_bgcolor='lightgray')
            fig.show()
    else:
        for ind in sector:
            fig = px.scatter(plot_df[plot_df['GICS Sector'] == ind], x = x, y = y, color = 'Year', hover_name= 'Symbol')
            fig.update_layout(title = 'Profit Pool of ' + ind,
                          xaxis_title = x,
                          yaxis_title = y,
                          title_x = 0.5,
                          plot_bgcolor='lightgray')
            fig.show()

# Process data
#GICS_df = pd.read_csv('gics.csv')
#latest_year = 2022
#pp_df = pd.DataFrame()
#for i in range(GICS_df.shape[0]):
#    sym = GICS_df.Symbol[i].replace('.', '-')
#    gics = GICS_df['GICS Sector'][i]
#    income = yf.Ticker(sym).income_stmt.T
#    income = income.reset_index()
#    income.rename(columns={'index': 'Date'}, inplace=True)
#    income.insert(0, 'Year', current_year - income.index)
#    income.insert(0, 'Symbol', sym)
#    income.insert(1, 'GICS Sector', gics)
#    pp_df = pd.concat([pp_df, income], ignore_index = True)
#    if (i % 50 == 0):
#        print(str(i) + " out of " + str(GICS_df.shape[0]))

# add profit margin
#pp_df['Profit Margin'] = pp_df['Net Income'] / pp_df['Total Revenue']
# save data
#pp_df.to_csv('pp_df.csv', index = False)
